feat: implement cli & esm.js
