+++ src/cli/args.js
@@ -1,5 +1,10 @@
 const parseArgs = () => {
-    // Write your code here 
+  console.log(
+    process.argv
+      .slice(2)
+      .map((e, i, a) => ((i + 1) % 2 === 0 ? `${a[i - 1]} is ${e}` : ''))
+      .join('')
+  );
 };
 
 parseArgs();
