+++ src/cli/env.js
@@ -1,5 +1,11 @@
 const parseEnv = () => {
-    // Write your code here 
+  console.log(
+    Object.entries(process.env)
+      .filter(([key, _]) => key.includes('RSS_'))
+      .map(([key, value]) => `${key}=${value}`)
+      .join('; ')
+      .trim()
+  );
 };
 
 parseEnv();
